pyproject.toml
[project]
name = "statelock-opt"
version = "0.1.0"
description = "Local eval-driven optimizer for StateLock-style memory systems"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

